+++ Assets/Scripts/UI/Leaderboards/PlayerGlobalLeader.cs
@@ -6,7 +6,7 @@ public class PlayerGlobalLeader : MonoBehaviour
     [SerializeField] private Text _playerName;
     [SerializeField] private Text _playerBalance;
 
-    public void Init(string name, int balance)
+    public void Init(string name, long balance)
     {
         _playerName.text = name;
         _playerBalance.text = balance.ToString() + " $";
